+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 - [code gen] Generate `@Origin` annotations for kotlin-inject-anvil and Metro code gen.
 - [code gen] Switch to `jakarta.inject` types for Dagger/Anvil code gen. This should have no source-breaking changes to users since this only affected generated code, but note that the square/anvil implementation may not support this in factory generation ([the KSP fork does](https://github.com/zacsweers/anvil)). If you need to only use javax annotations, use the `circuit.codegen.useJavaxOnly=true` KSP option.
 - [code gen] Drop KSP1 support.
+- [code gen] Fix not using named parameters with the `kotlin-inject` Anvil codegen mode.
 
 0.31.0
 ------